requirements.txt
streamlit
transformers
torch
Pillow
paddlepaddle
paddleocr
pdf2image
PyMuPDF